added methods to CRUD todo items

+++ TodoBackend.Context/Managers/TodoManager.cs
@@ -17,7 +17,16 @@ namespace TodoBackend.Context.Managers
         {
             using (TodoContext context = new TodoContext(contextOptions))
             {
-                context.CreateOrUpdateTodo(todoRecord);
+                Todo todo = context.CreateOrUpdateTodo(todoRecord);
+
+                if (todoRecord.TodoItems != null && todoRecord.TodoItems.Any())
+                {
+                    foreach (TodoItemRecord item in todoRecord.TodoItems)
+                    {
+                        context.CreateOrUpdateTodoItem(item, todo.Id);
+                    }
+                }
+
                 await context.SaveChangesAsync();
             }
         }
@@ -26,7 +35,7 @@ namespace TodoBackend.Context.Managers
         {
             using (TodoContext context = new TodoContext(contextOptions))
             {
-                Todo todo = await context.Todos.SingleOrDefaultAsync(x => x.Id == todoId);
+                Todo? todo = await context.Todos.SingleOrDefaultAsync(x => x.Id == todoId);
 
                 if (todo == null)
                 {
@@ -38,13 +47,33 @@ namespace TodoBackend.Context.Managers
             }
         }
 
+        public async Task UpdateTodoItemAsync(Guid todoId, Guid todoItemId, TodoItemRecord todoItemRecord)
+        {
+            using (TodoContext context = new TodoContext(contextOptions))
+            {
+                TodoItem? todoItem = await context.TodoItems.SingleOrDefaultAsync(x => x.TodoId == todoId && x.Id == todoItemId);
+
+                if (todoItem == null)
+                {
+                    return;
+                }
+
+                context.CreateOrUpdateTodoItem(todoItemRecord, todoId, todoItem);
+                await context.SaveChangesAsync();
+            }
+        }
+
         public async Task DeleteTodoAsync(Guid todoId)
         {
             using (TodoContext context = new TodoContext(contextOptions))
             {
-                Todo todo = await context.Todos.SingleOrDefaultAsync(x => x.Id == todoId);
-                context.Todos.Remove(todo);
-                await context.SaveChangesAsync();
+                Todo? todo = await context.Todos.SingleOrDefaultAsync(x => x.Id == todoId);
+
+                if (todo != null)
+                {
+                    context.Todos.Remove(todo);
+                    await context.SaveChangesAsync();
+                }
             }
         }
 
@@ -52,19 +81,23 @@ namespace TodoBackend.Context.Managers
         {
             using (TodoContext context = new TodoContext(contextOptions))
             {
-                TodoItem todoItem = await context.TodoItems.SingleOrDefaultAsync(x => x.Id == todoItemId);
-                context.TodoItems.Remove(todoItem);
-                await context.SaveChangesAsync();
+                TodoItem? todoItem = await context.TodoItems.SingleOrDefaultAsync(x => x.Id == todoItemId);
+                if (todoItem != null)
+                {
+                    context.TodoItems.Remove(todoItem);
+                    await context.SaveChangesAsync();
+                }
             }
         }
 
-        public async Task<List<Records.Outgoing.TodoRecord>> GetAllTodosAsync(int pageSize)
+        public async Task<List<Records.Outgoing.TodoRecord>> GetAllTodosAsync(int? pageNumber, int? pageSize)
         {
             using TodoContext context = new TodoContext(contextOptions);
 
             context.ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
 
             IQueryable<Records.Outgoing.TodoRecord> todosQuery = from todo in context.Todos
+                                                                 orderby todo.DateCreated descending
                                                                  select new Records.Outgoing.TodoRecord()
                                                                  {
                                                                      Id = todo.Id,
@@ -78,6 +111,15 @@ namespace TodoBackend.Context.Managers
                                                                          DueDate = todoItem.DueDate,
                                                                      }).ToList() : new List<Records.Outgoing.TodoItemRecord>()
                                                                  };
+
+
+            if (pageNumber.HasValue && pageSize.HasValue)
+            {
+                todosQuery = todosQuery
+                               .Skip(pageSize.Value * pageNumber.Value)
+                               .Take(pageSize.Value);
+            }
+
             return await todosQuery.ToListAsync();
         }
     }
